% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.R
\name{Weight}
\alias{Weight}
\title{Basic description}
\usage{
Weight(G)
}
\arguments{
\item{G}{a list of all groups}
}
\value{
the mean and weight matrix
}
\description{
Given the sampless, this function returns the mean and weight matrix.
}
\examples{
G=list()
set.seed(1)
n1=20;n2=40
N=n1+n2;
G[[1]]=c(1:n1);G[[2]]=c((n1+1):(n1+n2));
Weight(G)
}
